pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0",
]


build-backend = "setuptools.build_meta"

[project]
name = "playwright-computer-use"
version = "0.0.1"
authors = [
  { name="Marco Milanta", email="marco@invariantlabs.ai" },
]
description = "A package to connect Claude computer use with Playwright."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "anthropic",
    "python-dotenv",
    "playwright",
    "Pillow",
    "invariant-sdk",
    "python-dotenv",
]
[project.urls]
Homepage = "https://github.com/invariantlabs-ai/playwright-computer-use"
Issues = "https://github.com/invariantlabs-ai/playwright-computer-use/issues"


[tool.setuptools.package-data]
"playwright_computer_use" = ["assets/*.png"]


[tool.ruff]
select = ["D", "G"]  # Include docstring rules (D) and Google-style docstring rules (G)
ignore = ["G004"]          # List rules to ignore, if any
extend-ignore = []   # Optionally extend the ignore list

# Optionally, specify additional settings for Ruff
line-length = 88  # Adjust line length if needed

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
files = "src" 